+++ src/us/kbase/workspace/WorkspaceServer.java
@@ -104,7 +104,7 @@ public class WorkspaceServer extends JsonServerServlet {
 	//TODO INIT timestamps for startup script
 	//TODO DOCS workspace glossary
 
-	private static final String VER = "0.11.5";
+	private static final String VER = "0.12.0-dev1";
 	private static final String GIT = "https://github.com/kbase/workspace_deluxe";
 
 	private static final long MAX_RPC_PACKAGE_SIZE = 1005000000;